perf(test): breng de publicatie-stream-E2E's onder in één applicatie-instantie (#199)
* perf(test): breng de publicatie-stream-E2E's onder in één applicatie-instantie

De vier end-to-end-tests van de publicatie-stream hadden elk een eigen
resource-manager met eigen stub-servers. Quarkus start per testklasse met een
class-scoped resource een eigen applicatie-instantie plus database-container, dus
kostten vier tests die alleen verschillen in het antwoord van de aanmeld-stub ook
vier starts.

Ze staan nu als vier testmethodes in één klasse, op de gedeelde
DownstreamStubLifecycle. Het gedragsverschil (400, 500, eerst-500-dan-202) zit in
DownstreamHttpServer.statusVoorAanroep, dat van een constructor-parameter een
instelbare property is geworden; `reset()` zet het terug op 202 zodat een test het
gedrag van zijn voorganger niet erft.

Het retry-budget staat één keer in DownstreamStubLifecycle.MAX_POGINGEN; de
uitputtings-test leest die waarde in plaats van een eigen getal te hardcoderen.

Gemeten over de hele magazijn-module, lokaal, schone build:

| | vóór | na |
|---|---|---|
| applicatie-starts | 16 | 13 |
| boottijd | 75,3s | 67,2s |
| database-containers | 15 | 13 |
| moduletijd | 194s | 181s |

379 tests, ongewijzigd, allemaal groen. In CI-verhouding (~8,7s per start) scheelt
dit ongeveer 26s op de shard die berichtenmagazijn draait.


* perf(test): voeg de twee circuit-breaker-tests samen tot één instantie (#201)

* perf(test): voeg de twee circuit-breaker-tests samen tot één instantie

Beide tests hadden een eigen leeg profiel, puur om een schone applicatie-instantie af
te dwingen — breaker-state is application-scoped en blijft na een trip nog 5 seconden
open. Twee lege profielen kostten twee starts plus twee database-containers voor
dezelfde isolatie-eis; het zijn nu drie testmethodes in één klasse met één profiel.

De volgorde is functioneel en staat vast met @Order. De opslagtest gaat over HTTP en
valideert dus vóór het opslaan: draait die ná de Profiel-storingstest, dan fast-failt
hij op het inmiddels open validatiecircuit en wordt hij groen zonder het opslagcircuit
ooit te raken. Daarom opslag eerst, dan de weigering (die niets opent), en als laatste
de storing.

Gemeten over de hele magazijn-module, lokaal, schone build:

| | vóór | na |
|---|---|---|
| applicatie-starts | 16 | 15 |
| boottijd | 75,3s | 68,4s |
| database-containers | 15 | 14 |
| moduletijd | 194s | 186s |

379 tests, ongewijzigd, allemaal groen.


* perf(ci): verdeel de tests over twee machines (#195)

De modules draaiden achter elkaar in één reactor: 8:05, waarvan berichtenmagazijn
alleen al 4:38. Daarmee was de testcontrole met 433-548s de traagste van de keten en
bepaalde hij de wachttijd per wijziging — fuzzing zakte met #190 naar 316-376s.

Nu draaien berichtenmagazijn en de rest als aparte matrix-shard op een eigen runner.
Die ene module blijft de vloer (~5:15 inclusief de gedeelde bibliotheken), maar de
andere ~3:30 loopt ernaast in plaats van erna.

Beide shards bouwen met `-am` ook hun upstream-modules, dus fbs-common wordt twee keer
getest (~35s dubbel). Bewust: een expliciete lijst upstream-modules rot stilzwijgend
zodra de afhankelijkheden verschuiven, en dat kost dekking in plaats van tijd.

De check-run heet nog steeds `test`: een aggregerende job bundelt het shard-resultaat,
want branch protection op main pint op de context `checks-test / test`. Die job plaatst
ook de coverage-comment, over de rapporten van alle shards samen — los per shard zou
elke shard zijn eigen halve comment plaatsen.

De wijzigingsdetectie bepaalt de shard-verdeling, zodat de bestaande demo-only-scoping
één shard blijft. Zijn drie uitgangen zijn tot één afsluitend blok samengevoegd; de
fail-safe (bestandenlijst niet op te halen → alles draaien) is ongewijzigd.


---------


---------
