+++ services/berichtenmagazijn/src/test/kotlin/nl/rijksoverheid/moz/fbs/berichtenmagazijn/publicatie/DownstreamStubLifecycle.kt
@@ -3,10 +3,16 @@ package nl.rijksoverheid.moz.fbs.berichtenmagazijn.publicatie
 import io.quarkus.test.common.QuarkusTestResourceLifecycleManager
 
 /**
- * Quarkus-resource-manager voor [PublicatieStreamE2ETest]. Start twee
+ * Quarkus-resource-manager voor alle end-to-end-tests van [PublicatieStream]. Start twee
  * [DownstreamHttpServer]-instanties (`aanmeld`, `notificatie`) op random
  * loopback-poorten en exposeert hun URLs als config-overrides
  * (`magazijn.publicatie.downstreams.<key>.url`).
+ *
+ * Eén manager voor alle stream-tests: Quarkus start per unieke combinatie van
+ * resource-managers en profiel een eigen applicatie-instantie mét eigen database-container.
+ * Vier tests met elk hun eigen manager kostten vier starts; het gedragsverschil tussen die
+ * tests (400, 500, eerst-500-dan-202) zit nu in [DownstreamHttpServer.statusVoorAanroep],
+ * dat elke test in zijn eigen opzet instelt.
  *
  * Lifecycle-manager is de Quarkus-canonieke manier om servers vóór de
  * applicatie-boot te starten en hun URLs in de SmallRye-config te injecteren.
@@ -30,7 +36,13 @@ class DownstreamStubLifecycle : QuarkusTestResourceLifecycleManager {
         return mapOf(
             "magazijn.publicatie.downstreams.aanmeld.url" to aanmeld.baseUrl,
             "magazijn.publicatie.downstreams.notificatie.url" to notificatie.baseUrl,
+            // Polling-interval 200ms — Quarkus clampt naar 1s, snelste optie.
             "magazijn.publicatie.polling.interval" to "200ms",
+            // Lage backoff zodat een retry binnen het test-window valt.
+            "magazijn.publicatie.downstreams.aanmeld.backoff.basis" to "PT0.05S",
+            "magazijn.publicatie.downstreams.aanmeld.max-pogingen" to MAX_POGINGEN.toString(),
+            "magazijn.publicatie.downstreams.notificatie.backoff.basis" to "PT0.05S",
+            "magazijn.publicatie.downstreams.notificatie.max-pogingen" to MAX_POGINGEN.toString(),
             "quarkus.scheduler.enabled" to "true",
         )
     }
@@ -42,6 +54,13 @@ class DownstreamStubLifecycle : QuarkusTestResourceLifecycleManager {
     }
 
     companion object {
+        /**
+         * Retry-budget per downstream, gedeeld door alle stream-tests. De uitputtings-test
+         * leest deze waarde in plaats van een eigen getal te hardcoderen, zodat één budget
+         * volstaat en de tests niet uiteen hoeven te lopen in configuratie.
+         */
+        const val MAX_POGINGEN = 3
+
         /**
          * Test-globale registry zodat testcode na startup de juiste server-
          * instanties terugkrijgt voor assertions. Niet thread-safe maar test
