+++ services/berichtenmagazijn/src/test/kotlin/nl/rijksoverheid/moz/fbs/berichtenmagazijn/publicatie/PublicatieStreamE2ETest.kt
@@ -1,9 +1,8 @@
 package nl.rijksoverheid.moz.fbs.berichtenmagazijn.publicatie
 
+import io.quarkus.narayana.jta.QuarkusTransaction
 import io.quarkus.test.common.QuarkusTestResource
 import io.quarkus.test.junit.QuarkusTest
-import io.quarkus.test.junit.QuarkusTestProfile
-import io.quarkus.test.junit.TestProfile
 import io.restassured.RestAssured.given
 import io.restassured.http.ContentType
 import jakarta.inject.Inject
@@ -11,75 +10,60 @@ import jakarta.transaction.Transactional
 import nl.rijksoverheid.moz.fbs.berichtenmagazijn.opslag.BerichtRepository
 import org.awaitility.Awaitility
 import org.awaitility.Durations
+import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import java.util.concurrent.TimeUnit
 
 /**
- * End-to-end test voor [PublicatieStream]:
+ * End-to-end-tests voor [PublicatieStream], over de volle keten:
  *  1. POST /api/v1/berichten met `publicatietijdstip=now()`
  *  2. Quarkus Scheduler polt elke 200ms (override via [DownstreamStubLifecycle])
- *  3. PublicatieStream claimt deliveries, bouwt CloudEvent, levert af aan twee
- *     embedded HTTP-servers (één per geconfigureerde downstream)
- *  4. Test asserteert dat beide servers binnen 5 seconden een POST hebben ontvangen
- *     met `Content-Type: application/cloudevents+json` en een JSON-body waarin
- *     `type=nl.rijksoverheid.fbs.bericht.gepubliceerd` voorkomt.
+ *  3. PublicatieStream claimt deliveries, bouwt CloudEvents en levert af aan de twee
+ *     embedded HTTP-servers uit [DownstreamStubLifecycle]
+ *  4. De test asserteert wat er bij de stubs en in de delivery-rijen terechtkomt
  *
- * Geen WireMock-dependency: [DownstreamHttpServer] gebruikt `com.sun.net.httpserver`
- * uit de JDK. Resource-lifecycle ([DownstreamStubLifecycle]) zet downstream-URLs
- * vóór Quarkus de config initialiseert — system-properties uit een TestProfile
- * komen daarvoor te laat aan in `magazijn.publicatie.downstreams.*`.
+ * Geen WireMock-dependency: [DownstreamHttpServer] gebruikt `com.sun.net.httpserver` uit de
+ * JDK. De resource-lifecycle zet downstream-URLs vóór Quarkus de config initialiseert —
+ * system-properties uit een TestProfile komen daarvoor te laat aan in
+ * `magazijn.publicatie.downstreams.*`.
+ *
+ * De vier paden (aflevering, 4xx-terminal, 5xx-uitputting, retry) staan bewust in één klasse.
+ * Een class-scoped test-resource dwingt per testklasse een eigen applicatie-instantie mét eigen
+ * database-container af; als losse klassen kostten deze vier tests vier starts terwijl ze alleen
+ * verschillen in wat de aanmeld-stub antwoordt. Dat gedrag zet elke test nu zelf, op de gedeelde
+ * server.
  */
 @QuarkusTest
 @QuarkusTestResource(value = DownstreamStubLifecycle::class, restrictToAnnotatedClass = true)
-@TestProfile(PublicatieStreamE2ETest.E2EProfile::class)
 class PublicatieStreamE2ETest {
-
-    /**
-     * Verschillende `TestProfile` per E2E-test forceert Quarkus om de applicatie
-     * te restarten in plaats van een gedeelde JVM-context te hergebruiken. Zonder
-     * dit hergebruikt Quarkus de profile van de vorige test, wat tot stale
-     * `magazijn.publicatie.downstreams.*`-config kan leiden (URL's van een eerder
-     * @QuarkusTestResource).
-     */
-    class E2EProfile : QuarkusTestProfile {
-        override fun getConfigOverrides(): Map<String, String> = mapOf(
-            "quarkus.scheduler.enabled" to "true",
-        )
-    }
 
     @Inject
     lateinit var berichten: BerichtRepository
 
+    @Inject
+    lateinit var deliveries: PublicatieDeliveryRepository
+
+    private val aanmeld: DownstreamHttpServer
+        get() = DownstreamStubLifecycle.server("aanmeld")
+
+    private val notificatie: DownstreamHttpServer
+        get() = DownstreamStubLifecycle.server("notificatie")
+
+    /** Reset zet ook het antwoordgedrag terug op 202; elke test stelt zijn eigen pad in. */
     @BeforeEach
     @Transactional
     fun clean() {
+        deliveries.deleteAll()
         berichten.deleteAll()
-        DownstreamStubLifecycle.server("aanmeld").reset()
-        DownstreamStubLifecycle.server("notificatie").reset()
+        aanmeld.reset()
+        notificatie.reset()
     }
 
     @Test
     fun `aangeleverd bericht wordt naar beide downstreams gepubliceerd binnen polling-window`() {
-        given()
-            .contentType(ContentType.JSON)
-            .body(
-                """
-                {
-                  "afzender": "00000001003214345000",
-                  "ontvanger": {"type": "BSN", "waarde": "999993653"},
-                  "onderwerp": "E2E publicatie",
-                  "inhoud": "Test inhoud voor publicatie stream"
-                }
-                """.trimIndent(),
-            )
-            .`when`().post("/api/v1/berichten")
-            .then()
-            .statusCode(201)
-
-        val aanmeld = DownstreamStubLifecycle.server("aanmeld")
-        val notificatie = DownstreamStubLifecycle.server("notificatie")
+        lever(onderwerp = "E2E publicatie", inhoud = "Test inhoud voor publicatie stream")
 
         // Polling-interval is 200ms; downstreams moeten binnen enkele rondes ontvangen.
         Awaitility.await()
@@ -100,4 +84,186 @@ class PublicatieStreamE2ETest {
             "Notificatie body bevat event-type niet: ${notificatie.bodies.firstOrNull()}",
         )
     }
+
+    /**
+     * Client-fout (400) is niet-herstelbaar, dus de delivery moet meteen `MISLUKT` worden na
+     * exact 1 poging — geen retry.
+     *
+     * **Niet `volgende_poging IS NULL`**: [PublicatieDeliveryEntity.markeerMislukt] houdt de
+     * oude waarde aan bij terminal MISLUKT (kolom `nullable = false`, sentinel-design). Status
+     * `MISLUKT` is daarom de terminale-marker, niet de volgende_poging-tijd.
+     */
+    @Test
+    fun `400 op Aanmeld leidt tot MISLUKT na 1 poging zonder retry`() {
+        aanmeld.statusVoorAanroep = { _ -> 400 }
+
+        lever(onderwerp = "4xx-pad", inhoud = "Inhoud")
+
+        // Wacht tot de aanmeld-stub minstens 1× geraakt is en de delivery
+        // terminal MISLUKT is. Polling-interval 200ms (Quarkus clampt naar 1s).
+        Awaitility.await()
+            .atMost(15, TimeUnit.SECONDS)
+            .pollInterval(Durations.ONE_HUNDRED_MILLISECONDS)
+            .untilAsserted {
+                assertTrue(
+                    aanmeld.aantalAanroepen >= 1,
+                    "verwacht >= 1 call op Aanmeld",
+                )
+
+                val rijen = transactioneelOphalen()
+                val aanmeldRij = rijen.firstOrNull { it.doel == "aanmeld" }
+                    ?: error("aanmeld-delivery niet gevonden in rijen=${rijen.map { it.doel }}")
+
+                assertEquals(DeliveryStatus.MISLUKT, aanmeldRij.status, "status moet MISLUKT zijn")
+                assertEquals(1, aanmeldRij.pogingen, "geen retry: pogingen moet 1 zijn na enkele 400")
+            }
+
+        // Defense-in-depth tegen scheduler-clamp-races: `during(2s).atMost(3s)`
+        // verifieert dat de assertie 2 seconden lang ONONDERBROKEN blijft slagen.
+        // Een failure halverwege fail't de test direct (Awaitility short-circuit
+        // op de eerste mismatch — sneller dan blocking sleep). Bewijst dus
+        // "geen retry op 4xx" gedurende ~2 polling-intervals.
+        val callsNaTerminal = aanmeld.aantalAanroepen
+
+        Awaitility.await()
+            .during(2, TimeUnit.SECONDS)
+            .atMost(3, TimeUnit.SECONDS)
+            .pollInterval(Durations.ONE_HUNDRED_MILLISECONDS)
+            .untilAsserted {
+                assertEquals(
+                    callsNaTerminal,
+                    aanmeld.aantalAanroepen,
+                    "aanmeld mag na MISLUKT geen extra calls meer krijgen",
+                )
+
+                val aanmeldRij = transactioneelOphalen().firstOrNull { it.doel == "aanmeld" }
+                    ?: error("aanmeld-delivery weg")
+
+                // Als de scheduler hier ondanks MISLUKT-status toch zou claimen +
+                // retryen, zou pogingen >= 2 worden — afgevangen ook als stub-call-
+                // count flaky is door clamp-timing.
+                assertEquals(1, aanmeldRij.pogingen, "MISLUKT-rij mag niet opnieuw geprobeerd zijn")
+                assertEquals(DeliveryStatus.MISLUKT, aanmeldRij.status, "status moet MISLUKT blijven")
+            }
+    }
+
+    /**
+     * Een downstream die blijft falen (500) moet worden geretryd tot het budget uit
+     * [DownstreamStubLifecycle.MAX_POGINGEN] en daarna terminal `MISLUKT` worden — niet
+     * eindeloos doorgaan. Borgt de integratie van "herstelbaar maar uitgeput → terminal", die
+     * los alleen op unit-niveau ([RetryBeleidTest], [PublicatieClaimVerwerkerEdgeCaseTest])
+     * gedekt was.
+     */
+    @Test
+    fun `aanhoudende 500 leidt tot MISLUKT na maxPogingen retries`() {
+        aanmeld.statusVoorAanroep = { _ -> 500 }
+
+        lever(onderwerp = "5xx-pad", inhoud = "Inhoud")
+
+        // 500 is herstelbaar → retry tot het budget bereikt is → terminal MISLUKT.
+        Awaitility.await()
+            .atMost(15, TimeUnit.SECONDS)
+            .pollInterval(Durations.ONE_HUNDRED_MILLISECONDS)
+            .untilAsserted {
+                val aanmeldRij = transactioneelOphalen().firstOrNull { it.doel == "aanmeld" }
+                    ?: error("aanmeld-delivery niet gevonden")
+
+                assertEquals(DeliveryStatus.MISLUKT, aanmeldRij.status, "status moet MISLUKT zijn na uitputting")
+                assertEquals(
+                    DownstreamStubLifecycle.MAX_POGINGEN,
+                    aanmeldRij.pogingen,
+                    "moet exact maxPogingen pogingen hebben gedaan",
+                )
+            }
+
+        // Na terminal MISLUKT: geen verdere claims/calls meer (status sluit re-claim uit).
+        val callsNaTerminal = aanmeld.aantalAanroepen
+
+        Awaitility.await()
+            .during(2, TimeUnit.SECONDS)
+            .atMost(3, TimeUnit.SECONDS)
+            .pollInterval(Durations.ONE_HUNDRED_MILLISECONDS)
+            .untilAsserted {
+                assertTrue(
+                    aanmeld.aantalAanroepen <= callsNaTerminal,
+                    "aanmeld mag na MISLUKT geen extra calls meer krijgen",
+                )
+
+                val aanmeldRij = transactioneelOphalen().firstOrNull { it.doel == "aanmeld" }
+                    ?: error("aanmeld-delivery weg")
+
+                assertEquals(
+                    DownstreamStubLifecycle.MAX_POGINGEN,
+                    aanmeldRij.pogingen,
+                    "MISLUKT-rij mag niet opnieuw geprobeerd zijn",
+                )
+                assertEquals(DeliveryStatus.MISLUKT, aanmeldRij.status, "status moet MISLUKT blijven")
+            }
+    }
+
+    /**
+     * Retry-pad: eerste delivery aan Aanmeld faalt (HTTP 500), tweede slaagt. Borgt:
+     *  - [DownstreamClient] mapt non-2xx naar [DownstreamResultaat.Mislukt]
+     *  - [PublicatieStream] roept [PublicatieClaimer.markeerMislukt] met een berekende
+     *    `volgende_poging` ([RetryBeleid.volgendePoging])
+     *  - Bij volgende pollronde wordt de delivery opnieuw geclaimd en succesvol afgeleverd
+     *  - Notificatie-stub krijgt onafhankelijk gewoon één event (per-downstream isolatie)
+     */
+    @Test
+    fun `eerste 500 op Aanmeld leidt tot retry en tweede poging slaagt`() {
+        aanmeld.statusVoorAanroep = { poging -> if (poging == 1) 500 else 202 }
+
+        lever(onderwerp = "Retry-pad", inhoud = "Inhoud")
+
+        // Aanmeld faalt eenmalig (500), daarna 202. Notificatie altijd 202.
+        // Polling-interval is 1s (Quarkus clampt sub-1s naar 1s); retry-backoff
+        // (basis 50ms) voegt nog ~100ms toe. Verwacht binnen 15s een succesvolle
+        // tweede aflevering.
+        Awaitility.await()
+            .atMost(15, TimeUnit.SECONDS)
+            .pollInterval(Durations.ONE_HUNDRED_MILLISECONDS)
+            .untilAsserted {
+                // Aanmeld minstens 2 calls (1 fout + 1 succes), Notificatie 1.
+                assertTrue(
+                    aanmeld.aantalAanroepen >= 2,
+                    "verwacht >= 2 calls op Aanmeld (eerste 500, tweede 202), kreeg ${aanmeld.aantalAanroepen}",
+                )
+                assertTrue(
+                    notificatie.aantalAanroepen >= 1,
+                    "verwacht >= 1 call op Notificatie",
+                )
+            }
+
+        // Beide pogingen moeten dezelfde CloudEvent-`id` hebben (deterministisch per
+        // (berichtId, doel)) — andere attributen mogen verschillen (`time` is per attempt).
+        val bodies = aanmeld.bodies
+        val idRegex = Regex("\"id\":\"([^\"]+)\"")
+        val firstId = idRegex.find(bodies.first())?.groupValues?.get(1)
+        val secondId = idRegex.find(bodies[1])?.groupValues?.get(1)
+
+        assertEquals(firstId, secondId, "retry moet identieke CloudEvent-id hebben (deterministisch)")
+    }
+
+    private fun lever(onderwerp: String, inhoud: String) {
+        given()
+            .contentType(ContentType.JSON)
+            .body(
+                """
+                {
+                  "afzender": "00000001003214345000",
+                  "ontvanger": {"type": "BSN", "waarde": "999993653"},
+                  "onderwerp": "$onderwerp",
+                  "inhoud": "$inhoud"
+                }
+                """.trimIndent(),
+            )
+            .`when`().post("/api/v1/berichten")
+            .then()
+            .statusCode(201)
+    }
+
+    private fun transactioneelOphalen(): List<PublicatieDeliveryEntity> =
+        QuarkusTransaction.requiringNew().call {
+            deliveries.listAll()
+        }
 }
